feat: add Vite support for exported web components (#13148)
This PR adds Vite support for exported web components making it possible to build and run them with Vite in development and production modes.

A brief overview of the differences between Webpack and Vite implementations: 

**Webpack:**

When there is any component marked as exported, in addition to `bundle-***.js`, Webpack will build an `export-***.js` chunk that contains `generated-flow-imports`. The Flow client will be generated by Flow as a separate script.

Once an exported web component is requested, `WebComponentBootstrapHandler` is the one that is in charge of generating a bootstrap script to initialize this web component.

The current Webpack implementation of `WebComponentBootstrapHandler` is based on `BootstrapHandler`. It dynamically builds an HTML bootstrap document injecting all the scripts required for the exported web component to work to the document's head. There is where it will inject the Flow client as a separate script and the `export-***.js` chunk, getting it from the chunk stats file.

Eventually, the handler serializes the head of the resulting HTML document into a set of javascript instructions and sends them to the browser as a result.

**Vite:**

When there is any component marked as exported, Vite will build `vaadin-web-component.ts` and inject it to `web-component.html`. It is worth noting that `vaadin-web-component.ts` includes both `generated-flow-imports` and the Flow client.

For Vite, the `WebComponentBootstrapHandler` logic was modified such that it instead takes `web-component.html` as an initial bootstrap document. The bootstrap document already contains the actual link to `vaadin-web-component.ts` which is  
 injected by Vite. All the other scripts required for the exported web component to work, such as the `@Push` annotation script, the handler injects dynamically. An advantage of this approach is that it doesn't require any chunk stats so that it works pretty much the same way as `IndexHtmlRequestHandler`.

Eventually, the handler serializes the head of the resulting HTML document into a set of javascript instructions and sends them to the browser as a result.

Fixes #12298