refactor: Lookup code generator from the current context (#1283)
By avoiding a static reference to EndpointCodeGenerator, we can avoid having two references to the generator. The code generator is a bean and the context always has a reference to it

Before this, the EndpointCodeGenerator instances would remain in the application context even after the static reference was nulled. Then subsequent tests with the same configuration requirements would reuse the old application context -> you had 2 beans in use at the same time: one from the context and a new one in the static field